pyproject.toml
[project]
name = "techpanel"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.7",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "paramiko>=4.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pygithub>=2.9.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
    "PyJWT>=2.10.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
    "beautifulsoup4>=4.12.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
addopts = ["--import-mode=importlib"]
